transcribe/requirements.txt
openai-whisper
tqdm
librosa
Cython
Cmake
pyannote.core[notebook]
torch
torchaudio
pydub
nemo_toolkit[all]
